requirements.txt
asgiref==3.4.0
crispy-bootstrap5==0.4
Django==3.2.5
django-cleanup==5.2.0
django-crispy-forms==1.12.0
djangorestframework==3.12.4
Pillow==8.3.0
pytz==2021.1
sqlparse==0.4.1
